requirements.txt
astroid==2.2.5
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.3
django-rest-swagger==2.2.0
djangorestframework==3.10.2
idna==2.8
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
pep8==1.7.1
pkg-resources==0.0.0
pycodestyle==2.5.0
pylint==2.3.1
pytz==2019.2
requests==2.22.0
simplejson==3.16.0
six==1.12.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.3
wrapt==1.11.2
